README instructions for added editors & generators
When adding an editor or generator to a project, also add instructions
for how to run it to the README, if it exists and has a Rugs section.

Move from using the `scalar` method on the path expression engine to
the `with` method.

Stop using template-like replacement text in sample Rugs.

Closes #24