Get all Rugs and new-style tests working
Convert DSL tests to Gherkin.  Most work done by TestsToTypeScript.
It is not perfect, there is much duplication and auto-generated code,
but all tests are passing.  Update editors that add tests to add both
the Gherkin features and TypeScript implementations.  Remove tests
that check for specific version of Rug.  Remove DSL tests.

In rug 0.17.0, the Then step names are parsed as regular expressions,
so simplify their names to avoid any characters that can be
interpreted as special in regular expressions.

Add tests for AddLocalEditor.

Move parameter and variable names to TypeScript standard camelCase.

Update to rug-0.16.0.

Closes #39 and #38