Fix generated event handler
Previously the event handler added by AddTypeScriptEventHandler would
only work with events that had a `name` property, which was a small
subset of all the event types.  The generated test also failed because
the event type was not updated.  This is all fixed now.  Correct per
tslint along the way.