Expose the IncrementVersion function as a libbit
So, I'm building this thing (and it's showing up my talks)
which is an editor that creates an editor to supply a feature
(very specific kind: a file, and its tests) to another project
that wants to copy it.

This is the output of that. I used it to copy IncrementVersion
into another project where I wanted a custom BumpVersion editor.